contract/requirements.txt
numpy
py-solc-x
eth-brownie==1.14.2
